Add the first phase of scalar UDF and TMUDF tests
This submittal added the first phase of scalar UDF and TMUDF tests.  This only
covers scalar UDF and the C++ implementation of TMUDF.  The java implementation
of TMUDF is not covered.

The tests are currently disabled due to several pending bug fixes and
implementation changes.  They will be enabled once the feature is completely
done.

Change-Id: I946400ac0a8aed4dd8ecdd7365196e33c0731a52